Gem to GemMono
Changes to Grid and GridCollection to use the GemMono reference instead
of the base Gem class when creating and storing the objects. This is so
that it can keep track of each gem's visual representation and the
functionality portion at the same time without the Gem class needing a
reference to its proper MonoBehaviour partner.

Updated UML to reflect these changes

Updated relevant Information classes, Grid.cs, and GridCollection.cs