Component based Board
Converted creation of non-monobehaviour objects to be done by
subscribing to a creation event. The monobehaviour subscribes to a
static event and creates an instance of itself everytime a non-mono is
created. It then adds itself to the list of components on the non-mono
as an IComponent. That way there is a double link, but it is not so
strict that is must be specifically a GemMono class that the Gem class
has a reference to.

I suppose this is some sort of dependancy injection.