Swipe Gems
Proper swipe functionality is now implemented into GemMono.cs

UML does not currently  match

Chnages need to be made to GridCollectionMono to allow for easier
selection of a row or column based on a better 'collision' volume. Goal
is to make it so you can drag around on a row or column to move them
rather than requiring a gem to be drug.