Match with Fade animation
Matches can now be found. When a match is found, an animation plays.
After this animation, gravity pulls gems down to the bottom. This is all
based on IEnumerator functions which .MoveNext() when the event
'onCombatUpdate' is Invoked from the combat manager.

What is left is making new gems appear in their place.