Some Cleanup
Moved the IComponent interface to its own folder
Added an IAttachable interface with extension methods to GetComponent
and GetComponents

Added a new animation to the gem matches
Sped up the swiping animations

Cleaned up some of the code on Enemy.cs and Attribute.cs.

Added  a propertydrawer for the Attribute class.