Input Replay System
You're reading that correctly. I started working on an input replay
system. The ultimate capture of play tester data.

Since we already had an input wrapper for everything this was pretty
easy. Takes input and saves it as a simple class. OnApplicationQuit it
serializes the data as a .json. If you want to play it back, check
replay. It then reads the .json and attempts to send the input back out
as it happened the first time.

Works well for now, but is 100% impossible to be perfect. No one can
make it perfect. It is just a representation of what it may have looked
like.

Will work on saving Random.Seed values to try to replicate the random
generation as well.