Serialized Random Generator
Added a serialized random number generation manager. Now when replaying
data you can set the random number generator to the previous seed and it
will actually work. The issue that was happening last time was the fact
that UnityEvents are not 100% going to call functions back the same way
they were subscribed to. This meant that sometimes the replay would work
for one turn but most likely not. Now that there is a random generator
per object, there is no issue with getting the 'next' random number
before you were supposed to. It will always be the same.

More testing is required but it seems like this is working now.