Stamina Values and World State not Saving
The state of all the nodes was not being saved to the player data to be
loaded up. After the player comes back from combat, the
StageSelectionManager will handle putting the new state of the world on
the player.

The Stamina was not being Capped properly. It was possible for stamina
to go beyond the maximun limit without reseting. Will reset now if the
value goes byond the limit.