art push
added everything for the explosion in a folder name "Explosion_files" in
the assets folder.

the prefab for the actual explosion is
"Assets>Prefabs>damage_sparks.prefab"

I re-adjusted the size of the "Gatlin_enemy.prefab"to fit the screen
better.

I added more light probes to "hallway.prefab"so characters can be
positioned more freely.

I am currently working on fixing the animations so the pannels on the
front of the guns arnt visible till the gun is fired.